backend/requirements.txt
fastapi
uvicorn
python-multipart
minio
psycopg2-binary==2.9.7
sqlalchemy #sqlalchemy==2.0.20
pandas
pyarrow
#fire
spark-rapids-user-tools==23.10.0
slowapi==0.1.9
aiohttp==3.11.18
asyncpg==0.29.0
aiocache==0.12.2
tenacity==8.2.3
aiodns==3.1.1  # For async DNS resolution
#cchardet==2.1.7  # Faster character detection for aiohttp